GlobalFunctions: kfApiRequest now returns an object instead of array
Some tools already do this in their own "get API" method because of
performance reasons and ease of use (always passed by reference,
especially with large data sets this can get messy with arrays).

Folding that enhancement back into the main kfApiRequest method
so that we can get rid of the ad-hoc implementations.